Split CLI integration tests into modules
